Remove redundant utils
